+++ winbolo/src/winbolonet/http.c
@@ -103,7 +103,7 @@ bool httpCreate() {
   }
 
   sprintf(hostString, "Host: %s\n\n", wbnHostString);
-  hostStringLen = strlen(hostString);
+  hostStringLen = (int) strlen(hostString);
   
   httpStarted = returnValue;
   return returnValue;
@@ -246,7 +246,7 @@ int httpSendData(SOCKET sock, BYTE *message, int len) {
   if (buff != NULL) {
     encodeLen = httpEncodeData(message, len, buff, (len * 3));
     if (encodeLen != -1) {
-      returnValue = send(sock, HTTP_SEND_HEADER, strlen(HTTP_SEND_HEADER), 0);
+      returnValue = (int) send(sock, HTTP_SEND_HEADER, strlen(HTTP_SEND_HEADER), 0);
       if (returnValue > 0) {
         returnValue = send(sock, buff, encodeLen, 0);
 //DEBUG
@@ -408,7 +408,7 @@ bool httpSendLogFile2(char *fileName, BYTE *key, bool wantFeedback, long fileLen
     return FALSE;
   }
 
-  srand(time(NULL));
+  srand((unsigned int) time(NULL));
   randNum = rand() % 10;
   memset(sKey, 0, 33);
   memcpy(sKey, key, WINBOLONET_KEY_LEN);
@@ -421,7 +421,7 @@ bool httpSendLogFile2(char *fileName, BYTE *key, bool wantFeedback, long fileLen
   }
 
   sprintf(contentLength, "--%s\r\nContent-Disposition: form-data; name=\"logfile\"; filename=\"log.dat\"\r\nContent-Type: application/octet-stream\r\n\r\n", boundry); 
-  length = strlen(contentLength) + fileLength + strlen(boundry) + 7;
+  length = ((long) strlen(contentLength)) + fileLength + strlen(boundry) + 7;
 
   sprintf(header, "%s%s HTTP/1.0\r\nContent-Type: multipart/form-data; boundary=%s\r\nContent-Length: %ld\r\n%s", HTTP_POST_HEADER, sKey, boundry, length, hostString);  
   sock = socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
@@ -435,9 +435,9 @@ bool httpSendLogFile2(char *fileName, BYTE *key, bool wantFeedback, long fileLen
     }
 
 
-    ret = send(sock, header, strlen(header), 0);
+    ret = send(sock, header, (int) strlen(header), 0);
     if (ret != -1) {
-      ret = send(sock, contentLength, strlen(contentLength), 0);
+      ret = send(sock, contentLength, (int) strlen(contentLength), 0);
     }
     while (fileCount < fileLength && ret != -1) {
       if (fileCount + 512 < fileLength) {
@@ -455,7 +455,7 @@ bool httpSendLogFile2(char *fileName, BYTE *key, bool wantFeedback, long fileLen
     /* Send end of content marker */
     if (ret != -1) {
       sprintf(header, "\r\n--%s--\r\n", boundry);
-      ret = send(sock, header, strlen(header), 0);
+      ret = send(sock, header, (int) strlen(header), 0);
     }
 
     /* Get data back */
